DSP-10 move to pages file structure
