+++ index.html
@@ -48,7 +48,7 @@
     <a id="Filter"></a>
     <h2>Filter/Search:</h2>
     <div class="ebaBlue">Find a Provider That Offers...</div>
-    <form action="search/index.html" method="get">
+    <form action="/pages/search/index.html" method="get">
         <br>Service:
         <select name="Service">
             <option value="any" selected>(please select)</option>
