+++ index.js
@@ -74,7 +74,7 @@ const regionColor = (locality) => {
 
 const localities = simplemaps_statemap_mapdata.state_specific;
 for (let locality in localities) {
-  localities[locality].url = "locality/index.html?id=" + locality;
+  localities[locality].url = "pages/locality/index.html?id=" + locality;
   localities[locality].color = regionColor(locality);
 }
 //connect to data file
@@ -136,7 +136,7 @@ lastUpdatedSpan.innerText = directoryUpdated;
 var providers = dspsXML.getElementsByTagName("Provider");
 
 var locationCounter = 0;
-var officeURL = "/provider/index.html?id=";
+var officeURL = "pages/provider/index.html?id=";
 //cycle through all providers and add office locations to map
 for (let i = 0; i < providers.length; i++) {
   //get info for each provider
@@ -284,7 +284,7 @@ for (let i = 0; i < CSUStructure.length; i++) {
   const csuUL = document.createElement("ul");
   for (j = 0; j < CSUStructure[i].CSUs.length; j++) {
     const csuLI = document.createElement("li");
-    csuLI.innerHTML = `<a href="csu/index.html?id=${CSUStructure[i].CSUs[j].slug}">${CSUStructure[i].CSUs[j].name}</a>`;
+    csuLI.innerHTML = `<a href="pages/csu/index.html?id=${CSUStructure[i].CSUs[j].slug}">${CSUStructure[i].CSUs[j].name}</a>`;
     csuUL.appendChild(csuLI);
   }
   regionListUL.appendChild(regionHeading);
@@ -307,20 +307,20 @@ for (let i = 0; i < sortedProviders.length; i++) {
   //seperate the ID from the name
   var providerInfo = sortedProviders[i].split("!"); //providerInfo[0] is name [1] is ID
   const providerLI = document.createElement("li");
-  providerLI.innerHTML = `<a href="/provider/index.html?id=${providerInfo[1]}">${providerInfo[0]}</a>`;
+  providerLI.innerHTML = `<a href="pages/provider/index.html?id=${providerInfo[1]}">${providerInfo[0]}</a>`;
   providerUL.appendChild(providerLI);
 }
 
 const serviceUL = document.getElementById("serviceList");
 for (let i = 0; i < availableServices.length; i++) {
   const serviceLI = document.createElement("li");
-  serviceLI.innerHTML = `<a href="service/index.html?id=${i}">${availableServices[i]}</a>`;
+  serviceLI.innerHTML = `<a href="pages/service/index.html?id=${i}">${availableServices[i]}</a>`;
   serviceUL.appendChild(serviceLI);
 }
 
 const languageUL = document.getElementById("languageList");
 for (let i = 0; i < allLanguagesArray.length; i++) {
   const languageLI = document.createElement("li");
-  languageLI.innerHTML = `<a href="language/index.html?id=${i}">${allLanguagesArray[i]}</a>`;
+  languageLI.innerHTML = `<a href="pages/language/index.html?id=${i}">${allLanguagesArray[i]}</a>`;
   languageUL.appendChild(languageLI);
 }
