+++ pages/csu/index.html
@@ -4,11 +4,11 @@
 <head>
     <meta http-equiv="content-type" content="text/html; charset=UTF-8">
     <meta http-equiv="Cache-control" content="no-cache">
-    <title>Services Offered</title>
-    <script src="../mainmapdata.js"></script>
-    <script src="../statemap.js"></script>
+    <title>CSU Map</title>
+    <script src="../../mainmapdata.js"></script>
+    <script src="../../statemap.js"></script>
     <script type="module" src="index.js"></script>
-    <link rel="stylesheet" type="text/css" href="../pdstyle.css">
+    <link rel="stylesheet" type="text/css" href="../../pdstyle.css">
     <script>
         (function (i, s, o, g, r, a, m) {
             i['GoogleAnalyticsObject'] = r; i[r] = i[r] || function () {
@@ -19,15 +19,26 @@
 
         ga('create', 'UA-103149868-1', 'auto');
         ga('send', 'pageview');
+
+    </script>
+    <script>
+
     </script>
 </head>
 
 <body>
-    <a href="../index.html">Return to main map</a>
-    <h2><span id="serviceNameText"></span></h2>
+    <h2><span id="csuName"></span></h2>
+    <a href="../../index.html">Return to Main Map</a>.
+    <br><br>
     <div id="coveragemap"></div>
-    <div class="ebaBlue">The following providers offer this service:</div>
-    <ul id="providerUL">
+    <br><br>
+    <a href="#Providers">Sorted by Provider</a>&nbsp;|&nbsp;<a href="#Services">Sorted by Service</a>
+    <a id="Providers"></a>
+    <h2>Providers Serving This Court Service Unit:</h2>
+    <ul id="providerList">
+    </ul><a id="Services"></a>
+    <h2>Services offered in this Court Service Unit:</h2>
+    <ul id="serviceList">
     </ul>
 </body>
