+++ pages/language/index.html
@@ -21,7 +21,7 @@
 </head>
 
 <body>
-  <a href="../index.html">Return to main map</a>
+  <a href="../../index.html">Return to main map</a>
   <h2><span name="langaugeText"></span></h2>
   <div class="ebaBlue">The following providers offer services in <span name="langaugeText"></span>:</div>
   <ul id="providerList">
