+++ pages/language/index.js
@@ -1,5 +1,5 @@
-import dspsXML from "../src/getXML.js";
-import { removeDuplicates } from "../src/utils.js";
+import dspsXML from "../../src/getXML.js";
+import { removeDuplicates } from "../../src/utils.js";
 
 const queryString = window.location.search;
 const urlParams = new URLSearchParams(queryString);
