+++ pages/locality/index.html
@@ -4,11 +4,11 @@
 <head>
     <meta http-equiv="content-type" content="text/html; charset=UTF-8">
     <meta http-equiv="Cache-control" content="no-cache">
-    <title>CSU Map</title>
-    <script src="../mainmapdata.js"></script>
-    <script src="../statemap.js"></script>
+    <title>Locality Map</title>
+    <script src="../../mainmapdata.js"></script>
+    <script src="../../statemap.js"></script>
     <script type="module" src="index.js"></script>
-    <link rel="stylesheet" type="text/css" href="../pdstyle.css">
+    <link rel="stylesheet" type="text/css" href="../../pdstyle.css">
     <script>
         (function (i, s, o, g, r, a, m) {
             i['GoogleAnalyticsObject'] = r; i[r] = i[r] || function () {
@@ -21,25 +21,18 @@
         ga('send', 'pageview');
 
     </script>
-    <script>
 
-    </script>
 </head>
 
 <body>
-    <h2><span id="csuName"></span></h2>
-    <a href="../index.html">Return to Main Map</a>.
+
+    <h2><span id="localityText"></span></h2>
+    <br><br>
+    <a href="../../index.html">Return to All Providers Map</a>.
     <br><br>
     <div id="coveragemap"></div>
-    <br><br>
-    <a href="#Providers">Sorted by Provider</a>&nbsp;|&nbsp;<a href="#Services">Sorted by Service</a>
-    <a id="Providers"></a>
-    <h2>Providers Serving This Court Service Unit:</h2>
-    <ul id="providerList">
-    </ul><a id="Services"></a>
-    <h2>Services offered in this Court Service Unit:</h2>
-    <ul id="serviceList">
-    </ul>
+    <h2>Providers Serving This Locality:</h2><br>
+    <ul id="providerList" style='list-style: none;'>
 </body>
 
 </html>
