+++ pages/provider/index.html
@@ -24,7 +24,7 @@
 </head>
 
 <body>
-    <a href="/index.html"><i>Return to Provider Map Directory</i></a>.
+    <a href="../../index.html"><i>Return to Provider Map Directory</i></a>.
     <table cellpadding="10">
         <tr>
             <td>
