+++ pages/search/index.html
@@ -5,9 +5,9 @@
     <meta http-equiv="content-type" content="text/html; charset=UTF-8">
     <meta http-equiv="Cache-control" content="no-cache">
     <title>Search</title>
-    <script src="../mainmapdata.js"></script>
+    <script src="../../mainmapdata.js"></script>
     <script type="module" src="index.js"></script>
-    <link rel="stylesheet" type="text/css" href="../pdstyle.css">
+    <link rel="stylesheet" type="text/css" href="../../pdstyle.css">
     <script>
         (function (i, s, o, g, r, a, m) {
             i['GoogleAnalyticsObject'] = r; i[r] = i[r] || function () {
@@ -24,7 +24,7 @@
 </head>
 
 <body>
-    <a href="../index.html">Return to Main Map</a>.<br>
+    <a href="../../index.html">Return to Main Map</a>.<br>
     <h2>Providers offering
         <u><span id="serviceTextSpan"></span></u>
         in <u><span id="locationTextSpan"></span></u>
