+++ pages/service/index.html
@@ -4,11 +4,11 @@
 <head>
     <meta http-equiv="content-type" content="text/html; charset=UTF-8">
     <meta http-equiv="Cache-control" content="no-cache">
-    <title>Locality Map</title>
-    <script src="../mainmapdata.js"></script>
-    <script src="../statemap.js"></script>
+    <title>Services Offered</title>
+    <script src="../../mainmapdata.js"></script>
+    <script src="../../statemap.js"></script>
     <script type="module" src="index.js"></script>
-    <link rel="stylesheet" type="text/css" href="../pdstyle.css">
+    <link rel="stylesheet" type="text/css" href="../../pdstyle.css">
     <script>
         (function (i, s, o, g, r, a, m) {
             i['GoogleAnalyticsObject'] = r; i[r] = i[r] || function () {
@@ -19,20 +19,16 @@
 
         ga('create', 'UA-103149868-1', 'auto');
         ga('send', 'pageview');
-
     </script>
-
 </head>
 
 <body>
-
-    <h2><span id="localityText"></span></h2>
-    <br><br>
-    <a href="../index.html">Return to All Providers Map</a>.
-    <br><br>
+    <a href="../../index.html">Return to main map</a>
+    <h2><span id="serviceNameText"></span></h2>
     <div id="coveragemap"></div>
-    <h2>Providers Serving This Locality:</h2><br>
-    <ul id="providerList" style='list-style: none;'>
+    <div class="ebaBlue">The following providers offer this service:</div>
+    <ul id="providerUL">
+    </ul>
 </body>
 
 </html>
